deltapos is calculated when it begins to slow down
Originally deltapos was being calculated with the original starting position, so when the robot began to slow down it would drop drastically and then smoothly slow down. Now, deltapos is calculated with the current raw enc val and the raw enc val that the robot was at when it started slowing down.